+++ lib/src/paged_datatable_filter_bar.dart
@@ -184,12 +184,15 @@ class _FiltersDialog<TKey extends Object, TResult extends Object>
                                 style: const TextStyle(
                                     fontWeight: FontWeight.bold)),
                             const SizedBox(height: 8),
-                            ...state.filters.entries.where((element) => element.value._filter.visible).map((entry) => Padding(
-                                  padding:
-                                      const EdgeInsets.symmetric(vertical: 6),
-                                  child: entry.value._filter
-                                      .buildPicker(context, entry.value),
-                                ))
+                            ...state.filters.entries
+                                .where(
+                                    (element) => element.value._filter.visible)
+                                .map((entry) => Padding(
+                                      padding: const EdgeInsets.symmetric(
+                                          vertical: 6),
+                                      child: entry.value._filter
+                                          .buildPicker(context, entry.value),
+                                    ))
                           ],
                         ),
                       ),