+++ lib/src/paged_datatable_menu.dart
@@ -36,7 +36,8 @@ void _showMenu(
     {required BuildContext context, required List<BaseFilterMenuItem> items}) {
   final RenderBox button = context.findRenderObject() as RenderBox;
   var offset = button.localToGlobal(Offset.zero);
-  var position = RelativeRect.fromLTRB(offset.dx + 10, offset.dy + button.size.height - 10, 0, 0);
+  var position = RelativeRect.fromLTRB(
+      offset.dx + 10, offset.dy + button.size.height - 10, 0, 0);
 
   // var rect = RelativeRect.fromLTRB(offset.dx + 10, offset.dy + size.height - 10, 0, 0);